[core] fix compiler warning in 32-bit build

+++ src/ck.c
@@ -371,7 +371,7 @@ ck_backtrace (FILE *fp)
         }
 
         fprintf(fp, "%.2u: [%.012lx] (+%04x) %s\n",
-                frame, (uintptr_t)ip, (unsigned int)offset, name);
+                frame,(long unsigned)(uintptr_t)ip,(unsigned int)offset,name);
     }
     if (0 == rc)
         return;
